11/24/23 - KE's Fixes
Looked it over and it seemed a missed a few tests in addition to the 11 I did yesterday.

(I copied and pasted most of the stuff from existing stuff so if you could be on the lookout for incorrect variable names (i.e.: inside of an instructor class, referring to "ta" objects) that would be greatly appreciated)

There are a few methods that were added at the end of the document that I don't understand - can someone clarify their purpose before I do them?